server side auth setup & tidy up client side request error handling

+++ server/server.py
@@ -1,10 +1,11 @@
 import sqlite3
-from fastapi import FastAPI, APIRouter, Request
+from fastapi import FastAPI, APIRouter, Request, Response, status
 from fastapi.responses import RedirectResponse, HTMLResponse
 from fastapi.templating import Jinja2Templates
 from typing import Optional
 from pydantic import BaseModel
 import time
+import re
 
 class NewURLAlias(BaseModel):
 	slug: str
@@ -14,6 +15,22 @@ class NewURLAlias(BaseModel):
 	meta_colour: Optional[str]
 
 not_sus_website = "https://wikipedia.org/"
+
+tokens = ["ligma1000"]
+
+async def get_clearance_level(req: Request) -> int:
+	if req.headers["X-Clearance"]:
+		clearance = req.headers["X-Clearance"]
+		level_1_auth_regex = re.compile("^(?:Nerd|Peasant)\. No (?:doumentation|papers) found\.(?: Reccomended treatment: .+\.)*$")
+		level_3_auth_regex = re.compile("^(?:Gigachad|Absolute lad)\. Proof: (.+)\.$")
+
+		level_3_auth_regex_match = level_3_auth_regex.match(clearance)
+		if level_3_auth_regex_match is not None:
+			if level_3_auth_regex_match.group(1) in tokens:
+				return 3
+		elif level_1_auth_regex.match(clearance) is not None:
+			return 1
+	return 0
 
 db = sqlite3.connect("datameridian.db")
 cursor = db.cursor()
@@ -29,7 +46,11 @@ api = APIRouter()
 # 	return {"error": False, "data": {"message": "Hello, world!"}}
 
 @api.get("/url-aliases")
-async def get_all_url_aliases():
+async def get_all_url_aliases(request: Request, response: Response):
+	level = await get_clearance_level(request)
+	if level < 3:
+		response.status_code = status.HTTP_406_NOT_ACCEPTABLE
+		return {"error": True, "detail": "You don't have clearance to get information about all the url aliases."}
 	try:
 		aliases = []
 		for alias in cursor.execute("SELECT * FROM url_aliases ORDER BY uses DESC"):
@@ -53,7 +74,11 @@ async def get_all_url_aliases():
 		return {"error": True}
 
 @api.get("/url-aliases/{slug}")
-async def get_url_alias(slug: str):
+async def get_url_alias(request: Request, response: Response, slug: str):
+	level = await get_clearance_level(request)
+	if level < 1:
+		response.status_code = status.HTTP_406_NOT_ACCEPTABLE
+		return {"error": True, "detail": "You don't have clearance to get information on a url alias."}
 	try:
 		cursor.execute("SELECT * FROM url_aliases WHERE slug=:slug", {"slug": slug})
 		alias = cursor.fetchone()
@@ -73,7 +98,11 @@ async def get_url_alias(slug: str):
 		return {"error": True}
 
 @api.post("/url-aliases")
-async def create_url_alias(alias: NewURLAlias):
+async def create_url_alias(request: Request, response: Response, alias: NewURLAlias):
+	level = await get_clearance_level(request)
+	if level < 3:
+		response.status_code = status.HTTP_406_NOT_ACCEPTABLE
+		return {"error": True, "detail": "You don't have clearance to create url aliases."}
 	try:
 		cursor.execute("""INSERT INTO url_aliases
 		(alias_slug, canonical_url, created_at, uses, meta_title, meta_description, meta_colour)
@@ -86,7 +115,11 @@ async def create_url_alias(alias: NewURLAlias):
 		return {"error": True}
 
 @api.patch("/url-aliases/{id}")
-async def update_url_alias(alias: NewURLAlias, id: int):
+async def update_url_alias(request: Request, response: Response, alias: NewURLAlias, id: int):
+	level = await get_clearance_level(request)
+	if level < 3:
+		response.status_code = status.HTTP_406_NOT_ACCEPTABLE
+		return {"error": True, "detail": "You don't have clearance to edit url aliases."}
 	try:
 		cursor.execute("""UPDATE url_aliases SET 
 		alias_slug=:slug, canonical_url=:canonical_url, meta_title=:meta_title, meta_description=:meta_description, meta_colour=:meta_colour
@@ -99,7 +132,11 @@ async def update_url_alias(alias: NewURLAlias, id: int):
 		return {"error": True}
 
 @api.delete("/url-aliases/{slug}")
-async def delete_url_alias(slug: str):
+async def delete_url_alias(request: Request, response: Response, slug: str):
+	level = await get_clearance_level(request)
+	if level < 3:
+		response.status_code = status.HTTP_406_NOT_ACCEPTABLE
+		return {"error": True, "detail": "You don't have clearance to delete url aliases."}
 	try:
 		cursor.execute("DELETE FROM url_aliases WHERE alias_slug=:slug", {"slug": slug})
 		db.commit()
